- Added new form for login/password recovery request/password reset - Finished password reset function - Removed Google Maps API key from fragment/js/jsdump.jsp to prevent unwanted access; you'll have to use your own key from now on. - Added payment method radio button to checkout form. - All price fields in database now uses int instead of decimal to enhance performance.

+++ src/main/java/com/webbanhang2/dao/OrderDetailDaoImpl.java
@@ -119,7 +119,7 @@ public class OrderDetailDaoImpl implements OrderDetailDao {
             od.setOrderId(rs.getString("order_id"));
             od.setProductId(rs.getString("product_id"));
             od.setQuantity(rs.getInt("quantity"));
-            od.setPrice(rs.getDouble("price"));
+            od.setPrice(rs.getInt("price"));
             return od;
         }
     }
